pyproject.toml
[tool.poetry]
name = "electro-market"
version = "0.1.0"
description = ""
authors = ["y"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
django-rest-framework = "^0.1.0"
psycopg2 = "^2.9.9"
psycopg-binary = "^3.1.13"
python-dotenv = "^1.0.0"
djangorestframework-simplejwt = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
